+++ src/components/layout/index.ts
@@ -1,3 +1,2 @@
 export * from './flex';
-export * from './spacer';
 export * from './layout';