+++ src/components/spacer/spacer.test.tsx
@@ -1,0 +1,7 @@
+import { Spacer } from '.';
+
+describe('Spacer', () => {
+  it('is truthy', () => {
+    expect(Spacer).toBeTruthy();
+  });
+});